Crate quizx

Modules

annealer

basic_rules

Basic ZX-calculus rules

circuit

decompose

extract

fscalar

gate

generate

graph

hash_graph

json

Json encoding for interoperability with pyzx and zxlang using the .qgraph format.

linalg

Matrices and linear algebra over F2

optimize_circuit

phase

Phase encoded as either rational or floating point number of half-turns.

random_graph

scalar_traits

simplify

tensor

util

vec_graph