quizx

Module basic_rules

Basic ZX-calculus rules

These rules always come in triples of functions. For a rule X, there is a function check_X(&g, ...) -> bool which checks whether a rule is applicable at the given vertex or vertices, X_unchecked(&mut g, ...) applies the rule without doing any checking, and X(&mut g, ...) -> bool is the composition of the first two.

Note calling X_unchecked is allowed to make unsound ZX-diagram transformations, or even panic, if check_X doesn’t return true.

Functions

boundary_pivot

A checked implementation of the rule

check_boundary_pivot

check_color_change

Check color_change_unchecked applies

check_gadget_fusion

check_gen_pivot

Check gen_pivot applies

check_gen_pivot_reduce

Check gen_pivot applies and at least one vertex is interior Pauli

check_local_comp

Check local_comp_unchecked applies

check_pi_copy

Check pi_copy_unchecked applies

check_pivot

Check pivot_unchecked applies

check_remove_id

Check remove_id_unchecked applies

check_remove_pair

check_remove_single

check_spider_fusion

Check spider_fusion_unchecked applies

color_change

A checked implementation of the rule

color_change_unchecked

Change the color of a Z or X spider

gadget_fusion

A checked implementation of the rule

gadget_fusion_unchecked

gen_pivot

A checked implementation of the rule

gen_pivot_unchecked

Generic version of the pivot rule

local_comp

A checked implementation of the rule

local_comp_unchecked

Apply a local complementation

pi_copy

A checked implementation of the rule

pi_copy_unchecked

Apply a pi-copy

pivot

A checked implementation of the rule

pivot_unchecked

Apply pivoting to a pair of vertices

remove_id

A checked implementation of the rule

remove_id_unchecked

Remove an arity-2 spider with phase 0

remove_pair

A checked implementation of the rule

remove_pair_unchecked

Remove an isolated Z or X vertex and add it as a global scalar

remove_single

A checked implementation of the rule

remove_single_unchecked

Remove an isolated Z or X vertex and add it as a global scalar

spider_fusion

A checked implementation of the rule

spider_fusion_unchecked

Apply spider fusion