quizx::fscalar

Struct FScalar

pub struct FScalar { /* private fields */ }

This is the main representation of scalars used in QuiZX. It is a wrapper around four f64 values, used to represent the coefficients in a complex number of the form:

a + b ω + c ω² + d ω³

where ω is the 4th root of -1, i.e. exp(i π/4).

When scalars come from a Clifford+T circuit or ZX diagram, i.e. where all phase angles are integer multiples of π/4, scalars are guaranteed to be of the form above, with rational coefficients of the form x/2^y, for integers x, y.

These correspond exactly to the signficand and exponent of a floating point number (see e.g. floats on Wikipedia), so as long as x and y for each coefficient are not too large to fit in an f64, FScalar can exactly represent any Clifford+T scalar.

Note that ω² = i, so for all other complex numbers, FScalar gives an approximate representation of that number as a + c ω² = a + i c. This allows easy conversions to and from Complex<f64> provided in the standard library.

Implementations

impl FScalar

pub fn dyadic(pow: i32, coeffs: [i64; 4]) -> Self

Constructs a scalar from a power of two and 4 integer coefficients. For coeffs := [a,b,c,d], the resulting scalar is: 2^pow * (a + b ω + c ω² + d ω³).

pub fn real(r: f64) -> Self

Constructs a scalar as a real number with the given float value

pub fn complex(re: f64, im: f64) -> Self

Constructs a scalar as a complex number re + im * i.

pub fn one_plus_phase(phase: impl Into<Phase>) -> Self

Constructs a scalar 1 + e^(i π * phase) for the given Phase

pub fn mul_sqrt2_pow(&mut self, p: i32)

Convience method for multipling by the given power of sqrt(2)

pub fn mul_phase(&mut self, phase: impl Into<Phase>)

Convience method for multipling by e^(i π * phase) for the given Phase

pub fn mul_one_plus_phase(&mut self, phase: impl Into<Phase>)

Convience method for multipling by 1 + e^(i π * phase) for the given Phase

pub fn conj(&self) -> Self

Returns the complex conjugate of the scalar

pub fn complex_value(&self) -> Complex<f64>

Converts FScalar into a Complex<f64>

pub fn exact_dyadic_form(&self) -> [(i64, i16); 4]

Returns an array of 4 pairs giving each coefficient as a pair of integers (a,p), meaning a * 2^p. This is convenient e.g. for the base-2 scientific notation used to output scalars.

pub fn exact_phase_and_sqrt2_pow(&self) -> Option<(Phase, i16)>

If the scalar is an exact representation of a number in the form exp(i k π / 4) * sqrt(2)^p, return Some((k/4, p)), otherwise return None.