quizx::graph

Trait GraphLike

pub trait GraphLike:
    Clone
    + Sized
    + Send
    + Sync
    + Debug {

    // Required methods
    fn new() -> Self;
    fn vindex(&self) -> V;
    fn num_vertices(&self) -> usize;
    fn num_edges(&self) -> usize;
    fn vertices(&self) -> VIter<'_> ;
    fn edges(&self) -> EIter<'_> ;
    fn inputs(&self) -> &Vec<V>;
    fn inputs_mut(&mut self) -> &mut Vec<V>;
    fn set_inputs(&mut self, inputs: Vec<V>);
    fn outputs(&self) -> &Vec<V>;
    fn outputs_mut(&mut self) -> &mut Vec<V>;
    fn set_outputs(&mut self, outputs: Vec<V>);
    fn add_vertex(&mut self, ty: VType) -> V;
    fn add_vertex_with_data(&mut self, d: VData) -> V;
    fn add_named_vertex_with_data(&mut self, v: V, d: VData) -> Result<(), &str>;
    fn remove_vertex(&mut self, v: V);
    fn add_edge_with_type(&mut self, s: V, t: V, ety: EType);
    fn remove_edge(&mut self, s: V, t: V);
    fn set_phase(&mut self, v: V, phase: impl Into<Phase>);
    fn phase(&self, v: V) -> Phase;
    fn set_vertex_type(&mut self, v: V, ty: VType);
    fn vertex_type(&self, v: V) -> VType;
    fn vertex_data(&self, v: V) -> VData;
    fn set_edge_type(&mut self, s: V, t: V, ety: EType);
    fn edge_type_opt(&self, s: V, t: V) -> Option<EType>;
    fn set_coord(&mut self, v: V, coord: impl Into<Coord>);
    fn coord(&self, v: V) -> Coord;
    fn set_qubit(&mut self, v: V, qubit: f64);
    fn qubit(&self, v: V) -> f64;
    fn set_row(&mut self, v: V, row: f64);
    fn row(&self, v: V) -> f64;
    fn neighbors(&self, v: V) -> NeighborIter<'_> ;
    fn incident_edges(&self, v: V) -> IncidentEdgeIter<'_> ;
    fn degree(&self, v: V) -> usize;
    fn scalar(&self) -> &FScalar;
    fn scalar_mut(&mut self) -> &mut FScalar;
    fn find_edge<F>(&self, f: F) -> Option<(V, V, EType)>
       where F: Fn(V, V, EType) -> bool;
    fn find_vertex<F>(&self, f: F) -> Option<V>
       where F: Fn(V) -> bool;
    fn contains_vertex(&self, v: V) -> bool;

    // Provided methods
    fn add_to_phase(&mut self, v: V, phase: impl Into<Phase>) { ... }
    fn add_edge(&mut self, s: V, t: V) { ... }
    fn edge_type(&self, s: V, t: V) -> EType { ... }
    fn connected(&self, v0: V, v1: V) -> bool { ... }
    fn toggle_edge_type(&mut self, v0: V, v1: V) { ... }
    fn vertex_vec(&self) -> Vec<V> { ... }
    fn edge_vec(&self) -> Vec<(V, V, EType)> { ... }
    fn neighbor_vec(&self, v: V) -> Vec<V> { ... }
    fn incident_edge_vec(&self, v: V) -> Vec<(V, EType)> { ... }
    fn x_to_z(&mut self) { ... }
    fn add_vertex_with_phase(&mut self, ty: VType, phase: impl Into<Phase>) -> V { ... }
    fn add_edge_smart(&mut self, s: V, t: V, ety: EType) { ... }
    fn plug_vertex(&mut self, v: V, b: BasisElem) { ... }
    fn plug_output(&mut self, i: usize, b: BasisElem) { ... }
    fn plug_input(&mut self, i: usize, b: BasisElem) { ... }
    fn plug_inputs(&mut self, plug: &[BasisElem]) { ... }
    fn plug_outputs(&mut self, plug: &[BasisElem]) { ... }
    fn append_graph(&mut self, other: &impl GraphLike) -> FxHashMap<V, V> { ... }
    fn plug(&mut self, other: &impl GraphLike) { ... }
    fn is_identity(&self) -> bool { ... }
    fn tcount(&self) -> usize { ... }
    fn to_dot(&self) -> String { ... }
    fn adjoint(&mut self) { ... }
    fn to_adjoint(&self) -> Self { ... }
    fn component_vertices(&self) -> Vec<FxHashSet<V>> { ... }
    fn subgraph_from_vertices(&self, verts: Vec<V>) -> Self { ... }
    fn depth(&self) -> f64 { ... }

}

Required Methods

fn new() -> Self

Initialise a new empty graph

fn vindex(&self) -> V

Next fresh vertex index

fn num_vertices(&self) -> usize

Number of vertices

fn num_edges(&self) -> usize

Number of edges

fn vertices(&self) -> VIter<'_>

Get iterator over all vertices

fn edges(&self) -> EIter<'_>

Get iterator over all edges

An “edge” is a triple (s, t, edge_type), where s <= t.

fn inputs(&self) -> &Vec<V>

List of boundary vertices which serve as inputs

fn inputs_mut(&mut self) -> &mut Vec<V>

Mutable list of boundary vertices which serve as inputs

fn set_inputs(&mut self, inputs: Vec<V>)

Set inputs for the graph

fn outputs(&self) -> &Vec<V>

List of boundary vertices which serve as outputs

fn outputs_mut(&mut self) -> &mut Vec<V>

Mutable list of boundary vertices which serve as outputs

fn set_outputs(&mut self, outputs: Vec<V>)

Set outputs for the graph

fn add_vertex(&mut self, ty: VType) -> V

Add a vertex with the given type

fn add_vertex_with_data(&mut self, d: VData) -> V

Add a vertex with the given VData struct

fn add_named_vertex_with_data(&mut self, v: V, d: VData) -> Result<(), &str>

Add a vertex with the given name and VData struct

Returns an error if vertex already exists

fn remove_vertex(&mut self, v: V)

Remove a vertex from a graph

Behavior is undefined if the vertex is not in the graph.

fn add_edge_with_type(&mut self, s: V, t: V, ety: EType)

Add an edge with the given type

Behaviour is undefined if an edge already exists between s and t.

fn remove_edge(&mut self, s: V, t: V)

Remove an edge from a graph

Behaviour is undefined if there is no edge between s and t.

fn set_phase(&mut self, v: V, phase: impl Into<Phase>)

Set the phase of a vertex

fn phase(&self, v: V) -> Phase

Returns the phase of vertex v

fn set_vertex_type(&mut self, v: V, ty: VType)

fn vertex_type(&self, v: V) -> VType

fn vertex_data(&self, v: V) -> VData

fn set_edge_type(&mut self, s: V, t: V, ety: EType)

fn edge_type_opt(&self, s: V, t: V) -> Option<EType>

fn set_coord(&mut self, v: V, coord: impl Into<Coord>)

fn coord(&self, v: V) -> Coord

fn set_qubit(&mut self, v: V, qubit: f64)

fn qubit(&self, v: V) -> f64

fn set_row(&mut self, v: V, row: f64)

fn row(&self, v: V) -> f64

fn neighbors(&self, v: V) -> NeighborIter<'_>

fn incident_edges(&self, v: V) -> IncidentEdgeIter<'_>

fn degree(&self, v: V) -> usize

fn scalar(&self) -> &FScalar

fn scalar_mut(&mut self) -> &mut FScalar

fn find_edge<F>(&self, f: F) -> Option<(V, V, EType)>
where F: Fn(V, V, EType) -> bool,

fn find_vertex<F>(&self, f: F) -> Option<V>
where F: Fn(V) -> bool,

fn contains_vertex(&self, v: V) -> bool

Provided Methods

fn add_to_phase(&mut self, v: V, phase: impl Into<Phase>)

Adds a value to the phase of a vertex

fn add_edge(&mut self, s: V, t: V)

fn edge_type(&self, s: V, t: V) -> EType

fn connected(&self, v0: V, v1: V) -> bool

fn toggle_edge_type(&mut self, v0: V, v1: V)

fn vertex_vec(&self) -> Vec<V>

fn edge_vec(&self) -> Vec<(V, V, EType)>

fn neighbor_vec(&self, v: V) -> Vec<V>

fn incident_edge_vec(&self, v: V) -> Vec<(V, EType)>

fn x_to_z(&mut self)

Convert all X spiders to Z with the colour-change rule

fn add_vertex_with_phase(&mut self, ty: VType, phase: impl Into<Phase>) -> V

Add a vertex to the graph with the given type and phase

fn add_edge_smart(&mut self, s: V, t: V, ety: EType)

Add an edge and simplify if necessary to remove parallel edges

The behaviour of this function depends on the type of source/target vertex as well as the type of the existing edge (if there is one).

fn plug_vertex(&mut self, v: V, b: BasisElem)

Replace a boundary vertex with the given basis element

Note this does not replace the vertex from the input/output list or do normalisation.

fn plug_output(&mut self, i: usize, b: BasisElem)

Plug the given basis vertex into the i-th output.

fn plug_input(&mut self, i: usize, b: BasisElem)

Plug the given basis vertex into the i-th input.

fn plug_inputs(&mut self, plug: &[BasisElem])

Plug the given list of normalised basis elements in as inputs, starting from the left

The list plug should be of length <= the number of inputs.

fn plug_outputs(&mut self, plug: &[BasisElem])

Plug the given list of normalised basis elements in as outputs, starting from the left

The list plug should of length <= the number of outputs.

fn append_graph(&mut self, other: &impl GraphLike) -> FxHashMap<V, V>

Appends the given graph to the current one, with fresh names.

The renaming map is returned. The scalars are multiplied, but the inputs/outputs of self are NOT updated.

fn plug(&mut self, other: &impl GraphLike)

Plug the given graph into the outputs and multiply scalars

Panics if the outputs of self are not the same length as the inputs of other.

fn is_identity(&self) -> bool

Checks if the given graph only consists of wires from the inputs to outputs (in order)

fn tcount(&self) -> usize

Return number of Z or X spiders with non-Clifford phase

fn to_dot(&self) -> String

Return a graphviz-friendly string representation of the graph

fn adjoint(&mut self)

Exchange inputs and outputs and reverse all phases

fn to_adjoint(&self) -> Self

Same as GraphLike::adjoint(), but return as a copy

fn component_vertices(&self) -> Vec<FxHashSet<V>>

Returns vertices in the components of g

fn subgraph_from_vertices(&self, verts: Vec<V>) -> Self

Returns the full subgraph containing the given vertices

fn depth(&self) -> f64

Returns max row of any vertex

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.