quizx

Module json

Json encoding for interoperability with pyzx and zxlang using the .qgraph format.

Examples

// Define a graph with 4 vertices and 3 edges.
let mut g = Graph::new();
let vs = vec![
    g.add_vertex(VType::B),
    g.add_vertex(VType::Z),
    g.add_vertex(VType::X),
    g.add_vertex(VType::B),
];
g.set_inputs(vec![vs[0]]);
g.set_outputs(vec![vs[3]]);
g.add_edge(vs[0], vs[1]);
g.add_edge_with_type(vs[1], vs[2], EType::H);
g.add_edge(vs[2], vs[3]);

// Encode the graph in qgraph format.
let json = quizx::json::encode_graph(&g).unwrap();

// Decode the graph from the qgraph string.
let g2 = quizx::json::decode_graph::<Graph>(&json).unwrap();

assert_eq!(g.to_tensorf(), g2.to_tensorf());

Structs

JsonGraph

The json-encoded format for pyzx and zxlang graphs.

JsonPhase

A phase, in half turns.

JsonScalar

Global scalars in a graph

PhaseOptions

A set of options for encoding and decoding phases.

Enums

JsonError

An error that can occur when decoding a json graph.

Functions

decode_graph

Reads a graph from its json-encoded representation.

encode_graph

Returns the json-encoded representation of a graph.

read_graph

Reads a graph from a json-encoded file.

write_graph

Writes the json-encoded representation of a graph to a file.