quizx::linalg

Struct Mat2

pub struct Mat2 { /* private fields */ }

A type for matrices over F2

Implementations

impl Mat2

pub fn new(d: Vec<Vec<u8>>) -> Mat2

pub fn build<F>(rows: usize, cols: usize, f: F) -> Mat2
where F: Fn(usize, usize) -> bool,

Build a matrix with the given number of rows and columns. Place a 1 wherever f(i,j) is true.

pub fn zeros(rows: usize, cols: usize) -> Mat2

A matrix full of zeros

pub fn ones(rows: usize, cols: usize) -> Mat2

A matrix full of ones

pub fn id(dim: usize) -> Mat2

The identity matrix of a given size

pub fn unit_vector(dim: usize, i: usize) -> Mat2

A column vector with a single 1 at the given index

pub fn num_rows(&self) -> usize

pub fn num_cols(&self) -> usize

pub fn transpose(&self) -> Mat2

Return the transpose as a copy

pub fn gauss(&mut self, full_reduce: bool) -> usize

pub fn gauss_x( &mut self, full_reduce: bool, blocksize: usize, x: &mut impl RowOps, ) -> usize

pub fn rank(&self) -> usize

pub fn inverse(&self) -> Option<Mat2>

pub fn row_weight(&self, i: usize) -> u8

Return the hamming weight of the given row

pub fn weight(&self) -> u8

Return the hamming weight of the whole matrix

pub fn unit_rows(&self) -> Vec<usize>

Return a list of rows which have a single 1