quizx::phase

Struct Phase

pub struct Phase { /* private fields */ }

A phase, expressed in half-turns and encoded as a rational number.

The phase is always normalized to be in the range (-1,1].

Implementations

impl Phase

pub fn new(r: impl Into<Rational64>) -> Self

Creates a new phase.

Normalizes the phase to be in the range (-1,1].

pub fn to_rational(&self) -> Rational64

Returns the phase as a rational number.

pub fn from_f64(f: f64) -> Self

Creates a new phase from a floating point number of half-turns.

Rounds the floating point number to a rational number and normalizes it to be in the range (-1,1].

pub fn to_f64(&self) -> f64

Returns the phase as a floating point number of half-turns.

pub fn normalize(&self) -> Phase

Normalizes the phase to be in the range (-1,1] by adding or subtracting multiples of 2.

pub fn is_clifford(&self) -> bool

Returns true if the phase is a multiple of 1/2.

pub fn is_proper_clifford(&self) -> bool

Returns true if the phase is either -1/2 or 1/2.

pub fn is_pauli(&self) -> bool

Returns true if the phase is 0 or 1.

pub fn is_t(&self) -> bool

Returns true if the phase a non-clifford multiple of 1/4.

pub fn limit_denominator(&self, max_denom: i64) -> Self

Approximate a phase’s fraction to a Rational64 number with a small denominator.

Panics

Panics if max_denom is 0.