quizx::tensor

Trait QubitOps

pub trait QubitOps<A: TensorElem> {
    // Required methods
    fn ident(q: usize) -> Self;
    fn delta(q: usize) -> Self;
    fn cphase(p: impl Into<Phase>, q: usize) -> Self;
    fn hadamard() -> Self;
    fn delta_at(&mut self, qs: &[usize]);
    fn cphase_at(&mut self, p: impl Into<Phase>, qs: &[usize]);
    fn hadamard_at(&mut self, i: usize);
    fn slice_qubit_mut(
        &mut self,
        q: usize,
    ) -> (ArrayViewMut<'_, A, IxDyn>, ArrayViewMut<'_, A, IxDyn>);
    fn plug_n_qubits(self, n: usize, other: &Tensor<A>) -> Tensor<A>;
}

Required Methods

fn ident(q: usize) -> Self

fn delta(q: usize) -> Self

fn cphase(p: impl Into<Phase>, q: usize) -> Self

fn hadamard() -> Self

fn delta_at(&mut self, qs: &[usize])

fn cphase_at(&mut self, p: impl Into<Phase>, qs: &[usize])

fn hadamard_at(&mut self, i: usize)

fn slice_qubit_mut( &mut self, q: usize, ) -> (ArrayViewMut<'_, A, IxDyn>, ArrayViewMut<'_, A, IxDyn>)

split into two non-overlapping pieces, where index q=0 and q=1

fn plug_n_qubits(self, n: usize, other: &Tensor<A>) -> Tensor<A>

contract the last n qubit indices with the first n qubits of other

panics if n is greater than the number of qubits of self or other.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.